move errors to its own file
